Add logo upload to restaurant

+++ app/Services/FileService.php
@@ -13,6 +13,10 @@ class FileService
 
     public function deleteFile($filePath)
     {
+        if (str_starts_with($filePath, self::getFilePath())) {
+            return;
+        }
+
         if (file_exists($filePath)) {
             unlink($filePath);
         }
